Rename to 'Courier'.
