+++ test/courier_tests.cpp
@@ -14,7 +14,7 @@
         EXPECT_STREQ(ExpectedOut.c_str(), buffer.str().c_str());                                   \
     }
 
-TEST(Courierr, Warning)
+TEST(Courier, Warning)
 {
     ClientClass client_object("My object");
     std::string expected_output {"[WARNING] ClientClass(My object) LibraryClass(My library "
@@ -22,7 +22,7 @@ TEST(Courierr, Warning)
     EXPECT_STDOUT(client_object.library_object.generate_warning();, expected_output)
 }
 
-TEST(Courierr, Error)
+TEST(Courier, Error)
 {
     ClientClass client_object("My object");
     std::string expected_output {"[ERROR] ClientClass(My object) LibraryClass(My library object): "
@@ -31,7 +31,7 @@ TEST(Courierr, Error)
                   , expected_output)
 }
 
-TEST(Courierr, MessageLevel)
+TEST(Courier, MessageLevel)
 {
     ClientClass client_object("My object");
 
@@ -56,9 +56,9 @@ TEST(Courierr, MessageLevel)
     EXPECT_STDOUT(client_object.library_object.generate_warning();, expected_output)
 }
 
-TEST(Courierr, Minimal)
+TEST(Courier, Minimal)
 {
-    class MinimalCourier : public Courierr::Courierr {
+    class MinimalCourier : public Courier::Courier {
         void receive_error(const std::string& message) override {}
         void receive_warning(const std::string& message) override {}
         void receive_info(const std::string& message) override {}
@@ -72,12 +72,12 @@ TEST(Courierr, Minimal)
     }
     catch (std::runtime_error& exception) {
         EXPECT_STREQ(exception.what(),
-                     "Courierr: Error not handled by derived class: \"LibraryClass(My library "
+                     "Courier: Error not handled by derived class: \"LibraryClass(My library "
                      "object): Something serious happened\"");
     }
 }
 
-TEST(Courierr, Default)
+TEST(Courier, Default)
 {
     // Test message generated by LibraryDefaultCourier
     LibraryClass new_library_object("My library object");
